Refactoring: feature 폴더 train/trainResult 로직 분리
