+++ src/feature/train/trainResult/MakeNoticeTimes.js
