+++ src/feature/train/trainResult/makeCardMargin.js
